Create 0394. Decode String.js
I solved the problem using recusion but you can also use iterative technique.
our objective is to first find the character ']' and then finding the character '[' right befor it. and then finally we can keep track of the change by pushing the characters inside the bracket with the multiple of the number infront of the openning brackate. the rest it repeating the above steps using recursive calls